--- src/routes/auth.ts
@@ -1,67 +1,0 @@
-import express, { Request, Response } from 'express';
-import bcrypt from 'bcryptjs';
-import jwt from 'jsonwebtoken';
-import User from '../model/User';
-import Login from '../model/Login';
-import { AuthenticatedRequest } from '../middleware/auth';
-
-const router = express.Router();
-
-router.post('/register', async (req: AuthenticatedRequest, res: Response) => {
-  try {
-    const { username, password } = req.body;
-
-    const existingUser = await User.findOne({ username });
-    if (existingUser) {
-      return res.status(400).json({ message: 'User already exists' });
-    }
-
-    const salt = await bcrypt.genSalt(10);
-    const hashedPassword = await bcrypt.hash(password, salt);
-
-    const newUser = new User({
-      username,
-      password: hashedPassword,
-    });
-
-    await newUser.save();
-
-    res.status(201).json({ message: 'User registered successfully' });
-  } catch (error) {
-    console.error(error);
-    res.status(500).json({ message: 'Internal server error' });
-  }
-});
-
-router.post('/login', async (req: Request, res: Response) => {
-  try {
-    const { username, password } = req.body;
-
-    const existingUser = await User.findOne({ username });
-    if (!existingUser) {
-      return res.status(400).json({ message: 'Invalid username or password' });
-    }
-
-    const isPasswordValid = await bcrypt.compare(password, existingUser.password);
-    if (!isPasswordValid) {
-      return res.status(400).json({ message: 'Invalid username or password' });
-    }
-
-    const token = jwt.sign({ userId: existingUser._id }, process.env.SECRET!);
-
-    const newLogin = new Login({
-      username,
-      token: token,
-    });
-
-    await newLogin.save();
-
-    res.json({ token });
-  } catch (error) {
-    console.error(error);
-    res.status(500).json({ message: 'Internal server error' });
-  }
-});
-
-
-export default router;
