+++ src/server.ts
@@ -1,8 +1,9 @@
 import express from 'express';
 import mongoose, { ConnectOptions } from 'mongoose';
 import dotenv from 'dotenv';
-import authRoutes from './routes/auth';
-import todoRoutes from './routes/todo';
+import path from 'path';
+import authRoutes from './routes/authRoutes';
+import todoRoutes from './routes/todoRoutes';
 
 dotenv.config();
 
@@ -20,9 +21,13 @@ mongoose
   .catch((error) => console.error('MongoDB connection error:', error));
 
 app.use(express.json());
+console.log(path.join(__dirname, 'openapi'));
+
+app.use('/openapi', express.static(path.join(__dirname, 'openapi')));
+
 app.use('/api/auth', authRoutes);
-app.use('/api/todo', todoRoutes);
+app.use('/api/todos', todoRoutes);
 
 app.use((req, res) => {
-  res.status(404).json({ message: 'Not found' });
+  res.status(404).json({ message: `Route '${req.url}' does not exists or parameter is not provided` });
 });
